requirements.txt
hanzidentifier
SpeechRecognition
SoundFile
librosa
pydub
pinyin
tqdm
numpy
pandas
matplotlib
IPython
translate
scikit-learn
requests
torchvision